feat: update choice a and choice b widget of the animal game screen
